services/compliance_service/requirements.txt
fastapi[standard]>=0.116.1
pytest>=8.4.1
openai>=1.99.6
pydantic>=2.11.7
opentelemetry-distro
tenacity>=9.1.2